# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AlgoViz algorithm visualization platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AlgorithmCategory =
  (typeof AlgorithmCategory)[keyof typeof AlgorithmCategory];

export const AlgorithmCategory = {
  sorting: "sorting",
  searching: "searching",
  graph: "graph",
  "dsa-pattern": "dsa-pattern",
  cipher: "cipher",
  "data-structure": "data-structure",
} as const;

export type AlgorithmDifficulty =
  (typeof AlgorithmDifficulty)[keyof typeof AlgorithmDifficulty];

export const AlgorithmDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Algorithm {
  id: string;
  name: string;
  category: AlgorithmCategory;
  difficulty: AlgorithmDifficulty;
  description: string;
  timeComplexity: string;
  spaceComplexity: string;
  pseudocode: string[];
  tags: string[];
}

export interface AlgorithmListResponse {
  algorithms: Algorithm[];
}

export interface SaveProgressRequest {
  sessionId: string;
  completedSteps: number;
  totalSteps: number;
  timeSpentSeconds: number;
}

export interface SaveProgressResponse {
  success: boolean;
  message: string;
}

export interface UserProgressItem {
  algorithmId: string;
  sessionId: string;
  completedSteps: number;
  totalSteps: number;
  timeSpentSeconds: number;
  lastVisited: string;
}

export interface UserProgressListResponse {
  progress: UserProgressItem[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}