# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for AlgoViz algorithm visualization platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: algorithms
    description: Algorithm operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /algorithms:
    get:
      operationId: listAlgorithms
      tags: [algorithms]
      summary: List all available algorithms
      responses:
        "200":
          description: List of algorithms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlgorithmListResponse"
  /algorithms/{id}:
    get:
      operationId: getAlgorithm
      tags: [algorithms]
      summary: Get algorithm details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Algorithm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Algorithm"
        "404":
          description: Algorithm not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /algorithms/{id}/progress:
    post:
      operationId: saveProgress
      tags: [algorithms]
      summary: Save user progress for an algorithm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProgressRequest"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveProgressResponse"
  /progress:
    get:
      operationId: getUserProgress
      tags: [algorithms]
      summary: Get all user progress
      responses:
        "200":
          description: User progress list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgressListResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Algorithm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [sorting, searching, graph, dsa-pattern, cipher, data-structure]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        description:
          type: string
        timeComplexity:
          type: string
        spaceComplexity:
          type: string
        pseudocode:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - category
        - difficulty
        - description
        - timeComplexity
        - spaceComplexity
        - pseudocode
        - tags
    AlgorithmListResponse:
      type: object
      properties:
        algorithms:
          type: array
          items:
            $ref: "#/components/schemas/Algorithm"
      required:
        - algorithms
    SaveProgressRequest:
      type: object
      properties:
        sessionId:
          type: string
        completedSteps:
          type: integer
        totalSteps:
          type: integer
        timeSpentSeconds:
          type: integer
      required:
        - sessionId
        - completedSteps
        - totalSteps
        - timeSpentSeconds
    SaveProgressResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    UserProgressListResponse:
      type: object
      properties:
        progress:
          type: array
          items:
            $ref: "#/components/schemas/UserProgressItem"
      required:
        - progress
    UserProgressItem:
      type: object
      properties:
        algorithmId:
          type: string
        sessionId:
          type: string
        completedSteps:
          type: integer
        totalSteps:
          type: integer
        timeSpentSeconds:
          type: integer
        lastVisited:
          type: string
          format: date-time
      required:
        - algorithmId
        - sessionId
        - completedSteps
        - totalSteps
        - timeSpentSeconds
        - lastVisited
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message