# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AlgoViz algorithm visualization platform
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all available algorithms
 */
export const ListAlgorithmsResponse = zod.object({
  algorithms: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      category: zod.enum([
        "sorting",
        "searching",
        "graph",
        "dsa-pattern",
        "cipher",
        "data-structure",
      ]),
      difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
      description: zod.string(),
      timeComplexity: zod.string(),
      spaceComplexity: zod.string(),
      pseudocode: zod.array(zod.string()),
      tags: zod.array(zod.string()),
    }),
  ),
});

/**
 * @summary Get algorithm details
 */
export const GetAlgorithmParams = zod.object({
  id: zod.coerce.string(),
});

export const GetAlgorithmResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  category: zod.enum([
    "sorting",
    "searching",
    "graph",
    "dsa-pattern",
    "cipher",
    "data-structure",
  ]),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  description: zod.string(),
  timeComplexity: zod.string(),
  spaceComplexity: zod.string(),
  pseudocode: zod.array(zod.string()),
  tags: zod.array(zod.string()),
});

/**
 * @summary Save user progress for an algorithm
 */
export const SaveProgressParams = zod.object({
  id: zod.coerce.string(),
});

export const SaveProgressBody = zod.object({
  sessionId: zod.string(),
  completedSteps: zod.number(),
  totalSteps: zod.number(),
  timeSpentSeconds: zod.number(),
});

export const SaveProgressResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all user progress
 */
export const GetUserProgressResponse = zod.object({
  progress: zod.array(
    zod.object({
      algorithmId: zod.string(),
      sessionId: zod.string(),
      completedSteps: zod.number(),
      totalSteps: zod.number(),
      timeSpentSeconds: zod.number(),
      lastVisited: zod.date(),
    }),
  ),
});