# lib/api-zod/src/generated/types/algorithm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AlgoViz algorithm visualization platform
 * OpenAPI spec version: 0.1.0
 */
import type { AlgorithmCategory } from "./algorithmCategory";
import type { AlgorithmDifficulty } from "./algorithmDifficulty";

export interface Algorithm {
  id: string;
  name: string;
  category: AlgorithmCategory;
  difficulty: AlgorithmDifficulty;
  description: string;
  timeComplexity: string;
  spaceComplexity: string;
  pseudocode: string[];
  tags: string[];
}