# lib/api-zod/src/generated/types/algorithmListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for AlgoViz algorithm visualization platform
 * OpenAPI spec version: 0.1.0
 */
import type { Algorithm } from "./algorithm";

export interface AlgorithmListResponse {
  algorithms: Algorithm[];
}